Create test_hive_catalog.py for hive integration test only
